buisness_agent/pyproject.toml
[project]
name = "business-analyst-agent"
version = "1.0.0"
description = "Агент бизнес-аналитик для Evolution AI Agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"